fix: legacy 제거
