+++ WuxiaMeditation/Observable/MeditationObservable.swift
@@ -7,43 +7,6 @@
 
 import SwiftUI
 
-enum MeditationState {
-    case notStarted
-    case preparing
-    case progressing
-}
-
-enum MeditationRange {
-    case smallMeditation
-    case bigMeditation
-    
-    var time: Int {
-        switch self {
-        case .smallMeditation:
-            5
-        case .bigMeditation:
-            10
-        }
-    }
-}
-
-enum BreathState {
-    case inhale
-    case exhale
-    case pause
-    
-    static let inhaleExhale: Int = 5
-    static let pauseGap: Int = 2
-    
-    var wuxiaDescription: String {
-        switch self {
-        case .inhale: "흡 吸"
-        case .exhale: "호 呼"
-        case .pause: ""
-        }
-    }
-}
-
 @Observable
 final class MeditationObservable {
     var meditationState: MeditationState = .notStarted
@@ -88,7 +51,7 @@ final class MeditationObservable {
 
 // EneryLevel
 extension MeditationObservable {
-    // TODO: 어떤 기능을 위한 함수이지
+    
     func checkWuxiaTimeChanged() {
         let newDate = Date.now
         if currentWuxiaTime != newDate.wuxiaTime {
